Page projet : Editeur de texte : Création d'une fonction dédié à la requête de blocage d'édition et renommage des variables et fonctions. Si déconnexion du premier utilisateur, ce dernier n'écrasera pas les modifications des utilisateurs suivants si reconnexion.

+++ functions.php
@@ -501,10 +501,18 @@ function yproject_save_edit_project() {
 	ypcf_debug_log( 'yproject_save_edit_project > value ('.$current_lang.') => ' . $_POST['value'], TRUE );
 
 	//Supprime la réservation de l'édition en cours
-	$campaign_id = $campaign_id = filter_input( INPUT_POST, 'id_campaign' );
+	$WDGuser_current = WDGUser::current();
+	$user_id = $WDGuser_current->wp_user->ID;
+	$campaign_id = filter_input( INPUT_POST, 'id_campaign' );
 	$property = filter_input( INPUT_POST, 'property' );
 	$meta_key = $property.'_add_value_reservation';
-	delete_post_meta( $campaign_id, $meta_key );
+	$meta_value = get_post_meta( $campaign_id, $meta_key, TRUE ); 
+
+	if ( !empty($meta_value) ) {
+	    if ( $meta_value[ 'user' ] == $user_id ) {			
+			delete_post_meta( $campaign_id, $meta_key );
+	    }
+	}
 	
 	switch ($_POST['property']) {
 		case "title":
